% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankr.isp-package.R
\docType{package}
\name{bankr.isp-package}
\alias{bankr.isp}
\alias{bankr.isp-package}
\title{bankr.isp: Reads and Tidy Bank Statements from Intesa San Paolo}
\description{
Reads and Tidy Bank Statements from Intesa San Paolo.
}
\author{
\strong{Maintainer}: Andrea Ranzato \email{andrea.ranzato01@gmail.com}

}
\keyword{internal}
